+++ device-bootloaders/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-16846] Removed 32-bit BAR size fields in DIRs.
 ### Fixed
 - [SW-17528] Fix the alignment of DIRs and use GCC primitives.
+- [SW-17394] Fixed voltage calculation for throttle up and down operations
 ### Security
 
 ## [0.12.0] - 2023-06-02